+++ src/BaseClient.php
@@ -25,7 +25,7 @@ use yii\httpclient\Client;
  * @property string $id Service id.
  * @property string $name Service name.
  * @property array $normalizeUserAttributeMap Normalize user attribute map.
- * @property array $requestOptions HTTP request options. This property is read-only.
+ * @property-read array $requestOptions HTTP request options. This property is read-only.
  * @property StateStorageInterface $stateStorage Stage storage. Note that the type of this property differs in
  * getter and setter. See [[getStateStorage()]] and [[setStateStorage()]] for details.
  * @property string $title Service title.