+++ src/OAuthToken.php
@@ -16,9 +16,9 @@ use yii\helpers\ArrayHelper;
  * @property int $expireDuration Token expiration duration. Note that the type of this property differs in
  * getter and setter. See [[getExpireDuration()]] and [[setExpireDuration()]] for details.
  * @property string $expireDurationParamKey Expire duration param key.
- * @property bool $isExpired Is token expired. This property is read-only.
- * @property bool $isValid Is token valid. This property is read-only.
- * @property array $params This property is read-only.
+ * @property-read bool $isExpired Is token expired. This property is read-only.
+ * @property-read bool $isValid Is token valid. This property is read-only.
+ * @property-read array $params This property is read-only.
  * @property string $token Token value.
  * @property string $tokenSecret Token secret value.
  *